requirements.txt
Django==1.7.1
Fabric==1.10.1
Pygments==2.0.2
amqp==1.4.6
anyjson==0.3.3
arrow==0.5.0
billiard==3.3.0.19
celery==3.1.17
dj-database-url==0.3.0
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-faker==0.2
django-getenv==1.3.1
django-pipeline==1.4.3
django-redis-cache==0.13.0
djangorestframework==3.0.3
ecdsa==0.13
fake-factory==0.2
futures==2.2.0
gunicorn==19.1.1
hiredis==0.1.5
honcho==0.6.4
httpie==0.9.0
kombu==3.0.24
paramiko==1.15.2
psycopg2==2.5.4
pycrypto==2.6.1
python-dateutil==2.4.0
pytz==2014.10
redis==2.10.3
requests==2.5.1
six==1.9.0
sqlparse==0.1.14
static3==0.5.1
wsgiref==0.1.2
